Blocks: Expose blocks scoped hooks
