+++ blocks/api/registration.js
@@ -6,13 +6,9 @@
 import { get, isFunction, some } from 'lodash';
 
 /**
- * WordPress dependencies
- */
-import { applyFilters } from '@wordpress/utils';
-
-/**
  * Internal dependencies
  */
+import { applyFilters } from '../hooks';
 import { getCategories } from './categories';
 
 /**
