+++ blocks/api/serializer.js
@@ -8,11 +8,11 @@ import { html as beautifyHtml } from 'js-beautify';
  * WordPress dependencies
  */
 import { Component, createElement, renderToString, cloneElement, Children } from '@wordpress/element';
-import { applyFilters } from '@wordpress/utils';
 
 /**
  * Internal dependencies
  */
+import { applyFilters } from '../hooks';
 import { getBlockType, getUnknownTypeHandlerName } from './registration';
 
 /**
