+++ blocks/hooks/index.js
@@ -1,7 +1,7 @@
 /**
  * WordPress dependencies
  */
-import { addFilter } from '@wordpress/utils';
+import { createHooks } from '@wordpress/hooks';
 
 /**
  * Internal dependencies
@@ -11,7 +11,27 @@ import customClassName from './custom-class-name';
 import generatedClassName from './generated-class-name';
 import matchers from './matchers';
 
-const hooks = { addFilter };
+const hooks = createHooks();
+
+const {
+	addFilter,
+	removeFilter,
+	removeAllFilters,
+	applyFilters,
+	doingFilter,
+	didFilter,
+	hasFilter,
+} = hooks;
+
+export {
+	addFilter,
+	removeFilter,
+	removeAllFilters,
+	applyFilters,
+	doingFilter,
+	didFilter,
+	hasFilter,
+};
 
 anchor( hooks );
 customClassName( hooks );
