+++ blocks/hooks/test/geenerated-class-name.js
@@ -6,7 +6,7 @@ import { noop } from 'lodash';
 /**
  * External dependencies
  */
-import createHooks from '@wordpress/hooks';
+import { createHooks } from '@wordpress/hooks';
 
 /**
  * Internal dependencies
