+++ package.json
@@ -12,7 +12,7 @@
   ],
   "dependencies": {
     "@wordpress/a11y": "0.1.0-beta.1",
-    "@wordpress/hooks": "0.1.0-beta.5",
+    "@wordpress/hooks": "1.0.1",
     "@wordpress/url": "0.1.0-beta.1",
     "classnames": "2.2.5",
     "clipboard": "1.7.1",
